Trying to get it so only reroute files that are moved in the from directory if they were originally created there and then didn't pass the filter.
That is, if Chrome downloads 1.tmp, which bypasses the filter,
and then moves it to file.ext, file.ext will be rerouted,
but if other.tmp already exists and is moved to file.ext, it won't be rerouted.

It's complicated, however, and I haven't gotten it to work yet.
I might have a much simpler solution, however.
reroute just moves the from directory to from.real,
symlinks to => from, so that files saved in from are actually saved in to.